requirements.txt
certifi==2025.6.15
charset-normalizer==3.4.2
dropbox==12.0.2
idna==3.10
ply==3.11
python-dotenv==1.1.0
requests==2.32.4
six==1.17.0
stone==3.3.1
urllib3==2.5.0
